backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
PyJWT==2.8.0
email-validator==2.0.0
Pillow==10.0.0
gunicorn==21.2.0